fix: seq레디스로 관리

+++ src/main/java/com/example/petapp/domain/chatting/model/dto/LastMessageInfoDto.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Builder
 public class LastMessageInfoDto {
 
-    private int lastSeq;
+    private Long lastSeq;
     private String lastMessage;
     private String lastMessageTime;
 }
